Rename to reveal purpose better
