--- ImageEditor/src/main/kotlin/org/wordpress/android/imageeditor/home/HomeViewModel.kt
@@ -1,5 +1,0 @@
-package org.wordpress.android.imageeditor.home
-
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel()
