+++ ImageEditor/src/main/kotlin/org/wordpress/android/imageeditor/preview/PreviewImageFragment.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.imageeditor.home
+package org.wordpress.android.imageeditor.preview
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -9,17 +9,17 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProviders
 import org.wordpress.android.imageeditor.R
 
-class HomeFragment : Fragment() {
-    private lateinit var homeViewModel: HomeViewModel
+class PreviewImageFragment : Fragment() {
+    private lateinit var viewModel: PreviewImageViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.home_fragment, container, false)
+    ): View? = inflater.inflate(R.layout.fragment_preview_image, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        homeViewModel = ViewModelProviders.of(this).get(HomeViewModel::class.java)
+        viewModel = ViewModelProviders.of(this).get(PreviewImageViewModel::class.java)
     }
 }
