+++ ImageEditor/src/main/kotlin/org/wordpress/android/imageeditor/preview/PreviewImageViewModel.kt
@@ -1,0 +1,5 @@
+package org.wordpress.android.imageeditor.preview
+
+import androidx.lifecycle.ViewModel
+
+class PreviewImageViewModel : ViewModel()
