+++ ImageEditor/src/main/res/layout/activity_edit_image.xml
