+++ ImageEditor/src/main/res/layout/fragment_preview_image.xml
