+++ ImageEditor/src/main/res/navigation/nav_graph.xml
@@ -3,11 +3,11 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    app:startDestination="@+id/home_dest">
+    app:startDestination="@+id/preview_dest">
     <fragment
-        android:id="@+id/home_dest"
-        android:name="org.wordpress.android.imageeditor.home.HomeFragment"
-        android:label="@string/home"
-        tools:layout="@layout/home_fragment">
+        android:id="@+id/preview_dest"
+        android:name="org.wordpress.android.imageeditor.preview.PreviewImageFragment"
+        android:label="@string/preview"
+        tools:layout="@layout/fragment_preview_image">
     </fragment>
 </navigation>
