+++ ImageEditor/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">ImageEditor</string>
-    <string name="home">Home</string>
+    <string name="preview">Preview</string>
     <string name="crop">Crop</string>
 </resources>
