--- ImageEditor/src/test/kotlin/org/wordpress/android/imageeditor/home/HomeViewModelTest.kt
@@ -1,17 +1,0 @@
-package org.wordpress.android.imageeditor.home
-
-import org.junit.Before
-import org.junit.Test
-
-class HomeViewModelTest {
-    // Class under test
-    private lateinit var viewModel: HomeViewModel
-
-    @Before
-    fun setUp() {
-        viewModel = HomeViewModel()
-    }
-
-    @Test
-    fun `sample test`() {} // TODO:
-}
