+++ ImageEditor/src/test/kotlin/org/wordpress/android/imageeditor/preview/PreviewImageViewModelTest.kt
@@ -1,0 +1,17 @@
+package org.wordpress.android.imageeditor.preview
+
+import org.junit.Before
+import org.junit.Test
+
+class PreviewImageViewModelTest {
+    // Class under test
+    private lateinit var viewModel: PreviewImageViewModel
+
+    @Before
+    fun setUp() {
+        viewModel = PreviewImageViewModel()
+    }
+
+    @Test
+    fun `sample test`() {} // TODO:
+}
